Completion of dashboard page, and with a navbar and acard element.
The dashboard page displays list of courses user could take in the next quarter, in the future and coursestaken in the past. This is displayed as a card, which has a text that shows the course id. In this version, course id is hardcoded, not dynamic. When the card element is hovered and clicked, there are different behaviors. On click, description of that course appears as a popup. Add a multiple of additional node pacakges in this version, including tailwindcss, which is dramatically simplifies the process of styling. Tailwindcss generates a css file for you, using the className tag provided in elements. react-boostrap is also used to import navbar. Component hierarchy goes: App,DashBoard, (Navbar,Classfield), from Classfield, CardList,Card.