User interface for course selection page. Includes a navbar and selectioncard.
The course seleciton page is very similar to the dashboard page. The only difference is that the cards (the SelectionCard component) is a checkbox. The styling for the SelectionCard is identical to the Card component seen on the dashboard page, but a css file in used to style instead of using tailwindcss utility package. Yet to add a confirm button.