In progress code of integrating ryan's firebase to frontend connection code.
Make to new file 'getClassData.js' which is a funciton that will be called in 'SelectionCardList.js'. 'SelectionCardList' is ryan's code equivalent of 'CardSelect.js' which holds the CheckboxList component. This commiit version compiles and displays a webpage. However, in the 'getClassData' function, the error message "Error getting documents:" is returned when trying to fetch information from firebase and in 'SelectionCardList', error messsge "Error updating document" is returned. I predict that there's an issue connecting to firebase.