Change dashboard page's future courses and past course section so that it reflects database.
The dashboard page is now dynamic. Depending on the user's completed status for each class, the dashboard page's future courses section will display cards of classes that the user has yet to take and the past courses section will display the classes user has taken in the past. This is in link with the the selection page, so when a user select or deselect a course from the selection page, this will immediately get reflected on the dashboard page. This one done by using Ryan's code of retriving data from firestore. Ryan's code will retrieve information stored in firebase as an array. Filter out that array using .filter() and set the state variables array completedClasses and unCompletedClasses to include the filtered array. This is then passed on to CardList component as a prop so that each CardList component called could display cards of classes with respective state arrays.