Added class category, footer, and fixed login page styling.
Using the data retrieved from firestore, implemented a logic that categorizes courses in the course selection page. (using array.filter). User can now select classes based on category such as lower div and upper div classes. The footer has empty links with a animated wave on top for visual aesthetics. Most functionalities are implemented. There is a small bug on the selection page where the checkbox that is supposed to be pre-ticked is not. YT's pi chart is still not implemented into this version.