Trying to fix preticked checkbox problem.
When a returned list of courses shows that a user has already taken a course, that corresponding checkbox should be preticked. But when attemtping to implement this, the checkbox state will not visually update until page is reloaded. This commit is essentially an empty commit that does not change anything.