Fixed the pre-ticked checkbox bug.
The bug was on the selection page. The expected behavior was that for courses that a user already took, the checkbox of that corresponding course should come pre ticked, using the retreieved data from firebase. However, this behavior broke when implemnting categories when displaying courses, since I needed to create 5 different arrays. This happened because the event function that handled on a event where a checkbox of a course is ticked or unticked took in a index of the entire classData array. When this array that included all courses was split into five for categorization purposes, the index and its corresponding course became ambiguous because there was five arrays. If element 1 of array1 called the handleClick function with argument 1, handleClick function will not know whether this index corresponds to element 1 of array1 or array2 or array3 etc. Therefore, there was need to pass an additional parameter to the handleClick function that specified the array we wanted to modify the completed field. With the two arguments of index and specified array, the corresponding course we wanted to modify was now not ambiguous. Modified the handleClick function (handleCheckboxChange in code) so that depending on the argument, change the target that we want to modify. This version has every major element working together including the login page, selection page, dashboard and their UIs.