Added FileUpload component and edited related files.
FileUploadButton is part of the dashboardPage. This allows users to upload transcipt files and firebase will store them into its storage. Placed in the very buttom section of dashboard. User is able to upload one file at a time, and an success message will appear when the user was able to successfully upload their selected file. The file name of the file they selected also appears at the bottom of the selection button. Made modification to App.js and firebase.js to incorperate this component and also use the getStorage method provided by firebase. Changed the category titles (Past Courses, Future Courses etc) to a sticky behavior so that the position is fixed until the next sticky appears.