Made a searchbar in MyClasses page that allows to serach for classes.
The searchbar receieves a string and a class element that matches that string with its classCode will display on the screen. There are three elements: The searchbar, the search buttton and the clear button. The search button activates a search through funciton that does a search through firebase database and retrieves the class that matches user input. Once a match is found, all matching class cards will be the only cards displayed on the page. The clear button will clear whatever input there is on the search bar and will reset the filtering, meaning that all class cards will be displayed once the clear button is pressed. Searching through the database was done by uisng array.filter function.