Made a searchbar in MyClasses page that allows to search for classes.
The searchbar receives a string and a class element that matches thata
string with its classCode will display on the screen. There are three
elements: The searchbar, the search button and the clear button. The
search button activates a search through function that does a serach
through firebase database and retireves the clas that matches user
input. Once a match is found, all mathcing class cards will be the only
cards displayed on the page. The clear button will clear whatever input
there is on the search bar and will reset the filtering, meaning that
all class cards will be displayed once the clear button is pressed.
Searching through the database was done by using array.filter function.